+++ Rollbar/Deploys/RollbarDeploysManager.m
@@ -3,6 +3,7 @@
 #include <sys/utsname.h>
 #import "NSJSONSerialization+Rollbar.h"
 #import "RollbarDeploysManager.h"
+#include "Deployment.h"
 
 #define IS_IOS7_OR_HIGHER (floor(NSFoundationVersionNumber) > NSFoundationVersionNumber_iOS_6_1)
 #define IS_MACOS10_10_OR_HIGHER (floor(NSFoundationVersionNumber) > NSFoundationVersionNumber10_10)
@@ -12,7 +13,12 @@
 @property (readwrite, retain) NSString *readAccessToken;
 @end
 
-@implementation RollbarDeploysManager
+@implementation RollbarDeploysManager {
+    NSMutableData *responseData;
+    NSObject<DeploymentRegistrationObserver> *_deploymentRegistrationObserver;
+    NSObject<DeploymentDetailsObserver> *_deploymentDetailsObserver;
+    NSObject<DeploymentDetailsPageObserver> *_deploymentDetailsPageObserver;
+}
 
 - (id)initWithWriteAccessToken:(NSString *)writeAccessToken
                readAccessToken:(NSString *)readAccessToken
@@ -59,7 +65,7 @@ deploymentRegistrationObserver:(NSObject<DeploymentRegistrationObserver>*)deploy
     NSString * const url =
         @"https://api.rollbar.com/api/1/deploy/";
     NSMutableDictionary *params =
-        [[NSMutableDictionary alloc] initWithDictionary:deployment.asJSONData];
+        [[NSMutableDictionary alloc] initWithDictionary:deployment.jsonFriendlyData];
     [params setObject:self.writeAccessToken
                forKey:@"access_token"
      ];