+++ RollbarFramework/RollbarFramework.h
@@ -18,13 +18,15 @@ FOUNDATION_EXPORT const unsigned char RollbarFrameworkVersionString[];
 #import <RollbarFramework/RollbarKSCrashReportSink.h>
 #import <RollbarFramework/RollbarTelemetry.h>
 #import <RollbarFramework/RollbarTelemetryType.h>
+#import <RollbarFramework/RollbarLog.h>
 
 // Deploys API
 #import <RollbarFramework/RollbarDeploysProtocol.h>
 #import <RollbarFramework/RollbarDeploysManager.h>
 
-#import <RollbarFramework/RollbarJSONFriendlyProtocol.h>
-#import <RollbarFramework/RollbarJSONFriendlyObject.h>
+//#import <RollbarFramework/RollbarJSONFriendlyProtocol.h>
+//#import <RollbarFramework/RollbarJSONFriendlyObject.h>
+#import <RollbarFramework/DeployApiCallOutcome.h>
 #import <RollbarFramework/DeployApiCallResult.h>
 #import <RollbarFramework/Deployment.h>
 #import <RollbarFramework/DeploymentDetails.h>