+++ CHANGELOG.md
@@ -24,6 +24,11 @@ The change log has moved to this repo's [GitHub Releases Page](https://github.co
     depending on what is more appropriate in each case.
 
 ## Release Notes
+
+**1.10.0**
+- feat: ref #206: Capture Log As Telemetry - Not working? Introducing RollbarLog(...)
+- feat: ref #217: Reimplement Deploys based on the DTOs 
+
 **1.9.1**
 - feat: resolve #215: Change client.os element back to client.ios to fix dSYMs application